requirements.txt
# Core LangChain & LangGraph
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Vector Store for Memory
chromadb>=0.5.0

# Structured Output
pydantic>=2.0.0

# Web UI
fastapi>=0.115.0
uvicorn>=0.30.0
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0

# Optional: For OpenAI/Anthropic support
# langchain-openai>=0.2.0
# langchain-anthropic>=0.2.0